Move withStyles + overrideable test to hoc-interop test
This will be a general issue with higher-order components.
They infer the type of the props from the call signature. Because it is
overloaded we only infer a single case
