+++ packages/material-ui/test/typescript/styles.spec.tsx
@@ -14,7 +14,7 @@ import {
 import { ThemeProvider } from '@material-ui/styles';
 import Button from '@material-ui/core/Button';
 import { blue } from '@material-ui/core/colors';
-import { StandardProps, ButtonBase } from '@material-ui/core';
+import { StandardProps } from '@material-ui/core';
 
 // Shared types for examples
 interface ComponentProps extends WithStyles<typeof styles> {
@@ -462,22 +462,4 @@ withStyles(theme =>
   const CorrectUsage = () => <StyledMyButton nonDefaulted="2" />;
   // Property 'nonDefaulted' is missing in type '{}'
   const MissingPropUsage = () => <StyledMyButton />; // $ExpectError
-}
-
-// https://github.com/mui-org/material-ui/issues/14586
-{
-  const styles = createStyles({
-    root: {
-      color: 'red',
-    },
-  });
-
-  const StyledButton = withStyles(styles)(Button);
-
-  // undesired; caused by https://github.com/Microsoft/TypeScript/issues/26591
-  <StyledButton component="a" />; // $ExpectError
-
-  // workaround
-  const UnsafeStyledButton = withStyles({ root: { color: 'ref' } })(Button) as typeof Button;
-  <UnsafeStyledButton component="a" />;
 }
